configured accounts app according to User functions and templates are rearranged
