+++ accounts/templates/accounts/login_register.html
@@ -49,7 +49,7 @@
 
           <div class="auth__action">
             <p>Haven't signed up yet?</p>
-            <a href="{%  url 'base:register'  %}" class="btn btn--link">Sign Up</a>
+            <a href="{%  url 'accounts:register'  %}" class="btn btn--link">Sign Up</a>
           </div>
         </div>
       </div>
@@ -97,7 +97,7 @@
 
           <div class="auth__action">
             <p>Already have an account?</p>
-            <a href="{%  url 'base:login'  %}" class="btn btn--link">Log In</a>
+            <a href="{%  url 'accounts:login'  %}" class="btn btn--link">Log In</a>
           </div>
         </div>
       </div>
