+++ accounts/templates/accounts/update_user.html
@@ -6,7 +6,7 @@
         <div class="layout__box">
             <div class="layout__boxHeader">
                 <div class="layout__boxTitle">
-                    <a href="{% url 'base:profile' request.user.id %}">
+                    <a href="{% url 'accounts:profile' request.user.id %}">
                         <svg version="1.1" xmlns="http://www.w3.org/2000/svg" width="32" height="32"
                             viewBox="0 0 32 32">
                             <title>arrow-left</title>
