+++ base/models.py
@@ -1,18 +1,10 @@
 from email.policy import default
 from django.db import models
-from django.contrib.auth.models import AbstractUser
+from django.contrib.auth import get_user_model
 
 # Create your models here.
 
-class User(AbstractUser):
-    name = models.CharField(max_length=200, null=True, blank=True)
-    email = models.EmailField(unique=True, null = True)
-    bio = models.TextField(null=True, blank=True)
-    avatar = models.ImageField(null=True, default='avatar.svg')
-
-    USERNAME_FIELD = 'email'
-    REQUIRED_FIELDS = []
-
+User = get_user_model()
 
 class Topic(models.Model):
     name = models.CharField(max_length=200)
