+++ base/views.py
@@ -4,9 +4,9 @@ from django.http import HttpResponse
 from django.shortcuts import  get_object_or_404, redirect, render
 from django.contrib.auth.decorators import login_required
 from django.contrib import messages
-from django.contrib.auth import authenticate, login, logout
-from base.models import Room, Topic , Message, User
-from base.forms import RoomForm, UserForm, CustomUserCreationForm
+from base.models import Room, Topic , Message
+from accounts.models import User
+from base.forms import RoomForm
 # Create your views here.
 
 # rooms = [
@@ -38,15 +38,6 @@ def home(request):
 
 def about(request):
     return render(request, 'base/about.html')
-
-def profile_view(request, pk):
-    user = get_object_or_404(User, id=pk)
-    rooms = user.room_set.all()
-    topics_count = Topic.objects.count()
-    room_messages = user.message_set.all()
-    topics = Topic.objects.all()[:5]
-    context = {'user':user,'rooms':rooms, 'room_messages':room_messages, 'topics':topics, 'topics_count' : topics_count}
-    return render(request, 'base/profile.html', context)
 
 def room(request, pk):
     room = get_object_or_404(Room, id=pk)
@@ -134,58 +125,6 @@ def delete_room(request, pk):
     context = {'obj' :room}
     return render(request, 'base/forms/delete.html',context)
 
-def login_view(request):
-    page = 'login'
-    if request.user.is_authenticated:
-        return redirect('base:home')
-
-    if(request.method == 'POST'):
-        email = request.POST.get('email').lower()
-        password = request.POST.get('password')
-        
-        try:
-            temp_user = User.objects.get(email=email)
-        except:
-            messages.error(request, "User does not exists")
-            return render(request, 'base/forms/login_register.html', {})    
-            
-        
-        user = authenticate(request, email = email, password = password)
-        if(user is not None):
-            login(request, user)
-            messages.success(request, "Login success")
-            return redirect('base:home')
-        else:
-            messages.error(request, "username and password does not match")
-
-    context = {'page':page}
-    return render(request, 'base/forms/login_register.html', context)
-
-def logout_view(request):
-    logout(request)
-    messages.info(request, "Logout success")
-    return redirect('base:home')
-
-def register_view(request):
-    if request.user.is_authenticated:
-        return redirect('base:home')
-    page = 'register'
-    form = CustomUserCreationForm()
-    if request.method == 'POST':
-        form = CustomUserCreationForm(request.POST)
-        if(form.is_valid()):
-            user = form.save(commit=False)
-            user.username = user.username.lower()
-            user.email = user.email.lower()
-            user.save()
-            login(request, user)
-            return redirect('base:home')
-        else:
-            messages.error(request, "An error occured during registration")
-
-
-    context = {'page':page, 'form':form}
-    return render(request, 'base/forms/login_register.html', context)
 
 @login_required
 def delete_message(request, pk):
@@ -198,22 +137,6 @@ def delete_message(request, pk):
     return render(request, 'base/forms/delete.html', {'obj':msg})
 
 
-@login_required
-def update_user_view(request):
-
-    if(request.method == 'POST'):
-        form = UserForm(request.POST,request.FILES,  instance=request.user)
-        if form.is_valid():
-            form.save()
-            return redirect('base:profile', request.user.id)
-        else:
-            messages.warning(request, "something went wrong")
-
-    user = request.user
-    form = UserForm(instance=user)
-    context = {'form':form, 'user':user}
-
-    return render(request, 'base/forms/update_user.html', context)
 
 def topics_view(request):
